+++ src/external.cpp
@@ -41,6 +41,9 @@ void External::init (int new_max_var) {
   if (!max_var) {
     assert (e2i.empty ());
     e2i.push_back (0);
+    ext_units.push_back (0);
+    ext_units.push_back (0);
+    ext_flags.push_back (0);
     assert (internal->i2e.empty ());
     internal->i2e.push_back (0);
   } else {
@@ -53,6 +56,9 @@ void External::init (int new_max_var) {
     LOG ("mapping external %u to internal %u", eidx, iidx);
     assert (e2i.size () == eidx);
     e2i.push_back (iidx);
+    ext_units.push_back (0);
+    ext_units.push_back (0);
+    ext_flags.push_back (0);
     internal->i2e.push_back (eidx);
     assert (internal->i2e[iidx] == (int) eidx);
     assert (e2i[eidx] == (int) iidx);
@@ -64,6 +70,7 @@ void External::init (int new_max_var) {
       moltentab.resize (1 + (size_t) new_max_var, false);
   assert (iidx == (size_t) new_internal_max_var + 1);
   assert (eidx == (size_t) new_max_var + 1);
+  assert (ext_units.size () == (size_t) new_max_var * 2 + 2);
   max_var = new_max_var;
 }
 
@@ -140,13 +147,33 @@ void External::add (int elit) {
       (internal->opts.checkwitness || internal->opts.checkfailed))
     original.push_back (elit);
 
-  // The external literals of the new clause must be saved for later
-  // when the proof is printed during add_original_lit (0)
-  if (elit && internal->proof)
-    eclause.push_back (elit);
-
   const int ilit = internalize (elit);
   assert (!elit == !ilit);
+
+  // The external literals of the new clause must be saved for later
+  // when the proof is printed during add_original_lit (0)
+  if (elit && internal->proof) {
+    eclause.push_back (elit);
+    if (internal->opts.lrat && !internal->opts.lratexternal) {
+      // actually find unit of -elit (flips elit < 0)
+      unsigned eidx = (elit > 0) + 2u * (unsigned) abs (elit);
+      assert ((size_t) eidx < ext_units.size ());
+      const uint64_t id = ext_units[eidx];
+      bool added = ext_flags[abs (elit)];
+      if (id && !added) {
+        ext_flags[abs (elit)] = true;
+        internal->lrat_chain.push_back (id);
+      }
+    }
+  }
+
+  if (!elit && internal->proof && internal->opts.lrat &&
+      !internal->opts.lratexternal) {
+    for (const auto &elit : eclause) {
+      ext_flags[abs (elit)] = false;
+    }
+  }
+
   if (elit)
     LOG ("adding external %d as internal %d", elit, ilit);
   internal->add_original_lit (ilit);
@@ -345,8 +372,8 @@ void External::remove_observed_var (int elit) {
     int ilit = e2i[eidx]; // internalize (elit);
     internal->remove_observed_var (ilit);
 
-    melt (elit);
     is_observed[eidx] = false;
+    melt (elit);
     LOG ("unmarking %d as externally watched", eidx);
   }
 }
@@ -364,9 +391,9 @@ void External::reset_observed_vars () {
       int ilit = internalize (elit);
       internal->remove_observed_var (ilit);
       LOG ("unmarking %d as externally watched", eidx);
+      is_observed[eidx] = false;
       melt (elit);
     }
-    is_observed[eidx] = false;
   }
   internal->notified = 0;
   LOG ("reset notified counter to 0");
@@ -542,10 +569,16 @@ void External::melt (int elit) {
   unsigned &ref = frozentab[eidx];
   assert (ref > 0);
   if (ref < UINT_MAX) {
-    if (!--ref)
-      LOG ("external variable %d melted once and now completely melted",
-           eidx);
-    else
+    if (!--ref) {
+      if (observed (elit)) {
+        ref++;
+        LOG ("external variable %d is observed, can not be completely "
+             "molten",
+             eidx);
+      } else
+        LOG ("external variable %d melted once and now completely melted",
+             eidx);
+    } else
       LOG ("external variable %d melted once but remains frozen %u times",
            eidx, ref);
   } else