+++ src/tracer.cpp
@@ -4,9 +4,10 @@ namespace CaDiCaL {
 
 /*------------------------------------------------------------------------*/
 
-Tracer::Tracer (Internal *i, File *f, bool b, bool lrat, bool frat)
+Tracer::Tracer (Internal *i, File *f, bool b, bool lrat, bool frat,
+                bool veripb)
     : internal (i), file (f), binary (b), lrat (lrat), frat (frat),
-      added (0), deleted (0), latest_id (0) {
+      veripb (veripb), added (0), deleted (0), latest_id (0) {
   (void) internal;
   LOG ("TRACER new");
 }
@@ -268,10 +269,67 @@ void Tracer::drat_delete_clause (const vector<int> &clause) {
 
 /*------------------------------------------------------------------------*/
 
+void Tracer::veripb_finalize_proof (uint64_t conflict_id) {
+  if (file->closed () || !veripb)
+    return;
+  LOG ("TRACER veriPB tracing finalization of proof");
+  file->put ("output NONE\n");
+  file->put ("conclusion UNSAT : ");
+  file->put (conflict_id);
+  file->put (" \n");
+  file->put ("end pseudo-Boolean proof\n");
+}
+
+void Tracer::veripb_add_derived_clause (const vector<int> &clause,
+                                        const vector<uint64_t> &chain) {
+  LOG ("TRACER veriPB tracing addition of derived clause");
+  file->put ("pol ");
+  bool first = true;
+  for (auto p = chain.rbegin (); p != chain.rend (); p++) {
+    auto cid = *p;
+    if (first) {
+      first = false;
+      file->put (cid);
+    } else {
+      file->put (' ');
+      file->put (cid);
+      file->put (" + s");
+    }
+  }
+  file->put ("\n");
+  file->put ("e -1 ");
+  for (const auto &external_lit : clause) {
+    file->put ("1 ");
+    if (external_lit < 0)
+      file->put ('~');
+    file->put ('x');
+    file->put (abs (external_lit));
+    file->put (' ');
+  }
+  file->put (">= 1 ;\n");
+}
+
+void Tracer::veripb_begin_proof (uint64_t reserved_ids) {
+  LOG ("TRACER veriPB tracing start of proof");
+  file->put ("pseudo-Boolean proof version 2.0\n");
+  file->put ("f ");
+  file->put (reserved_ids);
+  file->put ("\n");
+}
+
+void Tracer::veripb_delete_clause (uint64_t id) {
+  LOG ("TRACER veriPB tracing deletion of clause");
+  file->put ("del id ");
+  file->put (id);
+  file->put ("\n");
+}
+
+/*------------------------------------------------------------------------*/
+
 void Tracer::add_original_clause (uint64_t id, const vector<int> &clause) {
   if (file->closed ())
     return;
-  if (frat)
+  else if (frat)
     frat_add_original_clause (id, clause);
 }
 
@@ -281,7 +339,7 @@ void Tracer::add_derived_clause (uint64_t id, const vector<int> &clause) {
   if (frat)
     frat_add_derived_clause (id, clause);
   else {
-    assert (!lrat); // TODO: there is some wierd bug with wierd options...
+    assert (!lrat && !veripb);
     drat_add_clause (clause);
   }
   added++;
@@ -291,7 +349,9 @@ void Tracer::add_derived_clause (uint64_t id, const vector<int> &clause,
                                  const vector<uint64_t> &chain) {
   if (file->closed ())
     return;
-  if (frat)
+  if (veripb)
+    veripb_add_derived_clause (clause, chain);
+  else if (frat)
     frat_add_derived_clause (id, clause, chain);
   else if (lrat)
     lrat_add_clause (id, clause, chain);
@@ -303,7 +363,9 @@ void Tracer::add_derived_clause (uint64_t id, const vector<int> &clause,
 void Tracer::delete_clause (uint64_t id, const vector<int> &clause) {
   if (file->closed ())
     return;
-  if (frat)
+  if (veripb)
+    veripb_delete_clause (id);
+  else if (frat)
     frat_delete_clause (id, clause);
   else if (lrat)
     lrat_delete_clause (id);
@@ -319,7 +381,13 @@ void Tracer::finalize_clause (uint64_t id, const vector<int> &clause) {
     frat_finalize_clause (id, clause);
 }
 
-void Tracer::set_first_id (uint64_t id) { latest_id = id; }
+void Tracer::set_first_id (uint64_t id) {
+  latest_id = id;
+  if (file->closed ())
+    return;
+  if (veripb)
+    veripb_begin_proof (id);
+}
 
 /*------------------------------------------------------------------------*/