+++ NEWS.md
@@ -1,3 +1,15 @@
+Version 1.7.1
+-------------
+
+ - Added support for VeriPB proofs (--lrat --lratveripb).
+
+ - Various fixes: LRAT proofs for constrain (which previously were
+   not traced correctly); internal-external mapping issues for LRAT
+   (worked for user propagator but now also in combination with LRAT);
+   further minor bug fixes.
+
+ - Added support for LRAT + external propagator in combination
+
 Version 1.7.0
 -------------